+++ src/main/java/com/airbus_cyber_security/graylog/wizard/config/rest/AlertWizardConfig.java
@@ -70,7 +70,7 @@ public abstract class AlertWizardConfig {
 						"",
 						0,
 						1,
-						0,
+						1,
 						"",
 						0,
 						"",